requirements.txt
python-docx
pyyaml
pyPDF2
